+++ lib/src/controls/form/text_form_box.dart
@@ -206,6 +206,7 @@ class TextFormBox extends FormField<String> {
                   scribbleEnabled: scribbleEnabled,
                   textDirection: textDirection,
                   selectionControls: selectionControls,
+                  initialValue: initialValue,
                 ),
               ),
             );