+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [next]
+## 4.1.1
 
 - Ensure acrylic is updated only if it's mounted ([#634](https://github.com/bdlukaa/fluent_ui/issues/634))
 - Ensure the provided `startYear` and `endYear` in `DateTime` are used properly ([#627](https://github.com/bdlukaa/fluent_ui/issues/627))